# specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Semiont API",
    "version": "0.1.0",
    "description": "Semantic Knowledge Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateResourceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "content": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/ContentFormat"
          },
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string"
          },
          "creationMethod": {
            "type": "string"
          },
          "sourceAnnotationId": {
            "type": "string"
          },
          "sourceResourceId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "content",
          "format",
          "entityTypes"
        ]
      },
      "CreateResourceResponse": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceDescriptor"
          },
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            }
          }
        },
        "required": [
          "resource",
          "annotations"
        ]
      },
...
      "GenerateResourceRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "Resource ID containing the annotation"
          },
          "title": {
            "type": "string",
            "description": "Custom title for generated resource"
          },
          "prompt": {
            "type": "string",
            "description": "Custom prompt for content generation"
          },
          "language": {
            "type": "string",
            "description": "Language locale (e.g., \"es\", \"fr\", \"ja\")"
          }
        },
        "required": [
          "resourceId"
        ]
      },
...
      "AnnotationContextResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          },
          "context": {
            "type": "object",
            "properties": {
              "before": {
                "type": "string"
              },
              "selected": {
                "type": "string"
              },
              "after": {
                "type": "string"
              }
            },
            "required": [
              "selected"
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceDescriptor"
          }
        },
        "required": [
          "annotation",
          "context",
          "resource"
        ]
      },
      "ContextualSummaryResponse": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "relevantFields": {
            "type": "object",
            "additionalProperties": {}
          },
          "context": {
            "type": "object",
            "properties": {
              "before": {
                "type": "string"
              },
              "selected": {
                "type": "string"
              },
              "after": {
                "type": "string"
              }
            },
            "required": [
              "selected"
            ]
          }
        },
        "required": [
          "summary",
          "relevantFields",
          "context"
        ]
      },
      "CreateAnnotationResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          }
        },
        "required": [
          "annotation"
        ]
      },
      "ListAnnotationsResponse": {
        "type": "object",
        "properties": {
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            }
          },
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "annotations",
          "total",
          "offset",
          "limit"
        ]
      },
...
      "GetAnnotationResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedResource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "annotation",
          "resource",
          "resolvedResource"
        ]
      },
...
      "AsyncJobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "complete",
              "failed",
              "cancelled"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "detection",
              "generation"
            ]
          },
          "created": {
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "status",
          "type",
          "created"
        ]
      },
      "CreateResourceFromAnnotationResponse": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceDescriptor"
          },
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          }
        },
        "required": [
          "resource",
          "annotation"
        ]
      },
...
      "DeleteAnnotationRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "format": "uri",
            "description": "Resource URI containing the annotation (required for O(1) Layer 3 lookup)"
          }
        },
        "required": [
          "resourceId"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/annotations/{id}": {
      "get": {
        "summary": "Get Annotation by URI (W3C Content Negotiation)",
        "description": "W3C-compliant globally resolvable annotation URI with content negotiation. Returns JSON-LD for API clients (Accept: application/ld+json or application/json). Redirects to frontend for browsers (Accept: text/html or browser User-Agent). Requires authentication and resourceId query parameter.",
        "tags": [
          "W3C URIs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Annotation ID"
          },
          {
            "name": "resourceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Resource ID or URI containing the annotation"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation representation (JSON-LD)",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnnotationResponse"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to frontend (for browsers)"
          },
          "400": {
            "description": "Missing resourceId parameter"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
    "/api/resources/{id}/annotations": {
      "get": {
        "summary": "Get Resource Annotations",
        "description": "Get all annotations (both highlights and references) in a resource",
        "tags": [
          "Resources",
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource annotations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResourceResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/annotations/{id}/create-resource": {
      "post": {
        "summary": "Create Resource from Annotation",
        "description": "Create a new resource from an annotation and resolve the annotation to it",
        "tags": [
          "Selections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource created and annotation resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResourceFromAnnotationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotations/{id}/generate-resource": {
      "post": {
        "summary": "Generate Resource (Job)",
        "description": "Create an async resource generation job from an annotation. Use GET /api/jobs/{jobId} to poll status. For real-time updates, use POST /api/annotations/{id}/generate-resource-stream instead.",
        "tags": [
          "Selections",
          "Resources",
          "Jobs",
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Annotation ID"
            },
            "required": true,
            "description": "Annotation ID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResourceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncJobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
    "/api/annotations/{id}/context": {
      "get": {
        "summary": "Get Annotation Context",
        "description": "Get the context around an annotation",
        "tags": [
          "Selections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "maximum": 5000,
              "default": 100
            },
            "required": false,
            "name": "contextBefore",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "maximum": 5000,
              "default": 100
            },
            "required": false,
            "name": "contextAfter",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationContextResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotations/{id}/summary": {
      "get": {
        "summary": "Get Contextual Summary",
        "description": "Get an AI-generated summary of the annotation in context",
        "tags": [
          "Selections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Contextual summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextualSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/annotations": {
      "post": {
        "summary": "Create Annotation",
        "description": "Create a new annotation/reference in a resource",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Annotation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnnotationResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Annotations",
        "description": "List all annotations for a resource (requires resourceId for O(1) Layer 3 lookup)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID to list annotations for"
            },
            "required": true,
            "description": "Resource ID to list annotations for",
            "name": "resourceId",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotations listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAnnotationsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/annotations/{id}": {
      "delete": {
        "summary": "Delete Annotation",
        "description": "Delete an annotation (requires resourceId in body for O(1) Layer 3 lookup)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Annotation deleted successfully"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
  },
...
}